Fix duplicate auth0Client params being sent
