+++ static/index.html
@@ -325,6 +325,7 @@
               console.log('Initializing using the factory function');
 
               createAuth0Client({
+                auth0Client: { name: 'my-client', version: 'some-version' },
                 domain: _self.domain,
                 client_id: _self.clientId,
                 cacheLocation: _self.useLocalStorage
